Optimize RSpec style and functionality
This change tightens up some specs to follow a general "describe ->
context -> it" pattern instead of "scenario" which is not so clear.
This change also removes the explicit inclusion of factory attributes
that are the default value, like { secret: false } for the patch
factory.  In other places, long automatically generated comments are
removed.