Handle missing OEmbed urls
A user on the production site removed their soundcloud audio preview on
soundcloud, but the link remained on VS and the OEmbed gem errored out
when trying to visit the patch detail page.  This change makes it so
that we handle missing OEmbed urls at the time they are rendered, hiding
missing urls on the patch detail page, and showing an error message on
index-level preview models.