Add Volca Keys images assets and style it onto the keys form
This was done by copying the image sizes, styles, and views used by the
Volca Bass form and then modifying them.  There may be some artifacts
left over from the Bass views that will need to be changed, for example
the form action in the form partial still points to the Bass endpoint.