Add '.bass' CSS class to bass show view
Without this class, the positional styles for the show bass view were
not applied and the knobs we incorrectly displayed in the top left
corner of the interface.  Adding this CSS class fixes that.  This is the
sort of thing that could be easily tracked with visual regression
testing but I had trouble finding a tool that would work.