Refactor the '.lit' CSS class
This change makes it so the color of lit and unlit knobs is its own
class on the knob DOM elements.  Before, it was coupled with the
'.small' class which made it inconsistent because it was an assumption.
Now the we are gonna support two different machines, it's better to be
more explicit so we have more control.