Render keys specific views when creating a patch fails
Bug: In the scenario where a user submits and invalid patch, and is
redirected back to the new patch form, that view was actually defaulting
to the bass patch form, and subsequently making changes and submitting
that errored out because it was trying to submit keys patch parameters
to create a new bass patch.

By adding 'keys/' to the view path, this specifies the keys form,
allowing users to make adjustments and successfully save the patch.