Add audio_sample_available field to keys patch model
This solves the issue where a user will set an audio sample that fits
the regex validations, but the oembed call to show the audio sample on
index and patch detail pages fails becuase it's not available.  This
pertains to soundcloud audio sample only.  Unavailable audio samples
also falsely boosts the patch quality score, affecting their prominence
on the index page.