Move #audio_sample_code to model level
This was formerly on the view model level because it was mainly just for
displaying the embed, but now we check whether the embed is available
during model validations, so it's been moved.  It's also moved into a
conern specific to audio samples because it's share by both the bass nad
keys models.