Add filter cutoff to volca bass simulator
This change makes it so that if you turn the cutoff knob on the
simulator, the cutoff of the sound generated will be affected.  This
also adds a data attribute to the knob object called "trueMidi" which is
a non-rounded midi value based on the rotation of the knob.  This
smooths out the changes in the generated sound and avoids "stepping".