Fix amp LFO behavior on bass emulator
BEFORE:
The lfo that modulated amplitude had a range of 1 to negative 1.  When
the volume is all the way up, the master amp's gain oscillated between 2
and 0.  Even though 2 is louder than the normal 1, it wasn't very
noticable.  However, if the volume was turned all the way down, the gain
was now oscillating between 1 and negative 1, resulting in an audible
signal even though the volume was all the way down.

AFTER:
In order to control the effect of the LFO on amp, a few things were put
in place.  The LFO oscillator is run through a wave shaper that adjusts
its range to have a range of 0 and negative 1.  This means that the
signal passed onto another amp can only turn it down, not up.  To make
it so this signal didn't override the main volume knob, a pre-amp was
added to accept all modulations in volume, before being passed to the
master amp.  This way, turning the knob all the way down will kill any
sound, regardless of whether any preAmp modulation is happening

NOTE:  It's uncertain whether the actual Volca Bass only modulates
amplification down.  It might modulate it up and down.  I will confirm
when I get my phsycial synth back and reference this behavior.