Map filter cutoff knob accurately using actual data and fancy math
It took a long time to find the right tool to come up with the equation
to map the filter cutoff curve.  Eventually I found a tool on
WolframAlpha that did the trick.  I gathered actual data from the synth
using the plot spectrum feature of audacity and plugged it into to this
calculator (https://tinyurl.com/y2qp9ebp) to come up with the equation
for the curve.