Store production database URI in different variable
This is in preparation for migrating the mLab hosted database over to
MongoDB Atlas.  When I remove the mLab add-on after migration, the
MONGODB_URI env variable will be deleted with it.  So by storing the
database URI in a new env variable, and pointing the production db
config to it, I don't risk losing database access when removing the
add-on.