Show 404 page for all non-conforming paths
this is to address the issue of 500 errors caused by bots requesting
paths that don't exist, like /apple-touch-icon.png, etc.  From now on,
all non-confirming paths will show the 404 page.  The format of the
response is forced to be html because if the request is for XML format,
rails tries to give back XML, which causes an EOFError unless we force
an html response.