Accept octave value from query string in bass emulator
This allows the main components of the emulator (interface, synth
engine, and the in-memory patch object) to accept a starting value for
octave via the query string.  It's programmed in a way to disallow
invalid values and default to octave 3 in case of an error.  More
parameters will be supported going forward so that users can effectively
save their patches by bookmarking URLs.