Add low level caching to homepage patch queries
This is an attempt to decrease calls to the database server and in turn,
decrease the operational costs.  I'm not 100% sure this is the best way
to do this, so I'm starting with the home page because the data does not
change often, and the patches that show up here aren't as critical as
the patch browse pages.