Accept cutoff value from query string in bass emulator
This change establishes a pattern where query string parameters must
match the id of their respective knob and must be provided in a midi
range value, like we store them in the database.  Then, we are
centralizing the calculation of the in-memory patch values in a series
of functions called "set{name_of_param}()" so that we can call them
dynamically.  This way we can also use the setter functions
initialization (reading query strings) and runtime (reactinf to knob
movements).

Note that octave has to conform to this new pattern and that has not
been done as part of this commit