Remove p5.js dependency
p5 was still included along with some other artifacts because this was
the initial solution for the emulator synth engine.  Now that we've
replaced the remaining usage of p5 (mainly key press events), we can
remove it altogether.  Hopefully this will enable us to test the
emulator functionality using rspec to the extent that we can: checking
knob rotation and if lights are lit based on query string params.