Test emulator query param patch setting functionality
This change makes it so we can test whether the interface reflects the
query params sent to the emulator.  This involves excluding the
AudioContext code from the test environment entirely so Poltergiest
doesn't choke on it.  This measure is hopefully temporary, until I can
find a web driver that can handle what we're actually doing.