Add osciallor note amps
This change adds in a set of amps between the oscillators and their
respective mute button amps.  As of writing, sometimes the mute amps (as
they're now called) were used for things like release times and note-of
functionality in addition to the mute buttons.  To make this more clear,
we're gonna have use the new note amps for note-off (and release)
purposes, and the mute amps will only be used for the mute buttons.

The mute amps should also be helpful in allow us to mute a single
osciallator instead of using the stop() function.  The stop() function
makes is have to create a new oscillator for each note, and swap that
node out in the signal path for each note played, which seems
inefficient and probably leads to memory problems longterm.

This change is with the hopes of being able to implement a sequencer
some time in the future.  It was looking very unreasonable because we
have to schedule notes to play and stop and with the current method of
swapping out the oscillator enturely for each note, that was not looking
very feasible.