Don't create a new oscillator for every new note.
This change implements the oscillator note amps to set their gain to 1
for a new note, and turn their gain to 0 when a note is released.
This was so we can use the same oscillators across many notes.

This should also help address memory issues when using the emulator for
longer periods of time.