Trigger filter decay/release on keyboardUp();
Self explanatory.  It's worth noting that if cancelAndHoldValue() is not
available, the audio parameter at the exacy moment is available using
the .value property: filter.frequency.value, so we can get it that way
without having to guess.  This method has also been applied to the
oscillator evnvelope instead of using the guess.  And by the way... the
guess was super duper precise anyway.