Improve bass emulator filter envelope a lot
Instead of modulating the filter frequency directly to create the filter
envelope, this change uses a constant source node to create envelope
values (between 0 and 1), and a gain node to amplify the envelope to
targer the filter's detune param.  This allows us to move the VCF cutoff
knob in the middle of a filter EG without reseting the filter frequency
as we hear it.

It turns out that two nodes can modulate the filter frequency at the
same time.  And that's what happening here.  The filterEgAmp and
ampLfoCutoff are both modulating the filter's detune at the same time.