Mute oscillators and envelope completely instead of 0.0001
This is so Firefox won't say "Playing" in the browser tab title area
even when it's silent.  Also I think Chrome considers the very very
quiet oscillations to be Auotplay functionality and I don't want it to
automatically block sound from the page.  We were using 0.0001 because
it's needed for exponentialRampToValueAtTime() but that's note being
used anymore.