Differentiate keyboardUp from stopping the note
There were two different keyboard up functions that did two different
things.  One determined whether to stop the note or change the note.
The other just stopped the note.  This renames the latter from
keyboardUp() to stopNote() because that's what it does.

This change also uses the time parameter instead of default values or 0.