Fix EG sequencer issue in Bass emulator
When the sequencer is playing, and the attack is turned up on the
envelope, the note will stop playing if the attack becomes longer than
the time to the next note.  This is because cancelScheduledValues() will
kill a ramp immediately even if the time given is after the ramp has
begun.

This fix only works for Chrome (and maybe others).  Firefox will not get
this fix because it does not implement cancelAndHoldValueAtTime().  So I
added a note so the user knows this.