Move some variables to Tone.Param objects
Some variables whose values were being used in real-time were moved to
Tone.Param types so they could be scheduled and play nicer with the
sequencer:
- notePlaying
- attackEndTime

Some existing audio params were also moved to the Tone.Param object
because their values were being used in real-time and there was an
opportunity to use Tone.Param.getValueAtTime() instead, which uses
linear interpolation (among other types) to calculate the value of a
param at a certain point on the timeline.  For example in the middle of
a ramp.

This change is experimental and is still being tested.  It seems the
sequencer on Firefox has timing issues and has a glitchy swing or
something.  One thing I haven't confirmed is whether
Tone.Param.setValueAtTime(time) will schedule the event against Tone's
clock or the master context clock.  There could be a potential for
timing issues there.