Tidy up stopNote EG code
Since we are using Tone.Param for envelopes now, Firefox can now support
cancelAndHoldAtTime because Tone monkey patched it.

This change also skips the rest of the stopNote function if the note is
already silent.

Some variables are mmoved so they are only instantiated if they are
needed.

These last two points presumably would help performance.