Throttle mousemove events
This change creates a throttle mechanism that will only fire the
turnKnob event 40 times a second instead of 125. This should help
performance on the bass emulator.  In addition, instead of also
listening to the document 'mousemove' event, the bass emulator code will
now be listening to the 'knobturn' which is dispatched directly from
form turnKnob.  So we don't have to throttle two things at the same time
and risk the throttle timers being staggered and acting weird.