Switch to chrome as a test driver
This change removes Poltergeist in favor of using chrome as the headless
browser for the test suite.  Poltergeist has been out of support for a
while and it would break on more modern Javascripts.  Some tests needed
to be adjusted because the driver interface changed a little or behavior
within the actual test browser was slightly different.  None of the
changes were super critical so adjustments to the tests were made.