Dockerize dev environment
This is the basic implementation of letting the app run in a dockerized
environment.

To build and run:
docker compose up --build

Then you need to attach the the vs container and create a first user
record in order for the homepage to load.  The plan is to do this
automatically for new containers / first volumes.

This change does not account for differences on the production
environment so it should not be pushed to prod.