Bind audio engine changes to input change, not interface interaction
This addresses the issue where if you click on a button on the bottom
row of emulator controls, and drag off of the button before mouseUp, the
underlying checkbox would not change, the light would not change, but
the audio engine would register the change, for example the VCO wave
would toggle.  That's because this counts as a click/tap javascript
event but it does not count as toggling the underlying control.

This change makes it so the audio engine changes are bound the the
changes of the underlying control, ensuring synchronization.