Add step record functionality
Now the user can click the step record button and the next notes they
input will be set as the sequencer notes.  A few things to note:
Clicking the step record button will stop the sequencer if it's already
playing.  Playing the sequence will exit step record mode if it's on.