Add skip step button to step record mode
On the actual synth, you have to hold down shift to enable and disable
step record mode.  Since that's uncommon to do on a GUI, I made the
entire record button the step record button and when it's active, a
smaller button to skip a step appears above it.  In short, a two button
solution instead of hold one, press one.

Skipping a step marks 'step mode' as false for the current step and
advances the value of stepModeIndex, which is basically a cursor used ot
edit the sequence without playing it.

Additional changes:
 - Add min-width of 320px to body of entire site.  No reason to support
   anything under this.
 - Add a .nolight class for interface buttons that don't toggle
   lit/unlit when they're pressed.  So far this is only permalink and
   skip-step.