Move synth params to its own file.
Part of an effort to clean up the main bass_emulator.es6 file and
separate it into smaller more managable chunks.

This moves the synth param object into it's own file.  The synth param
object is used as the canoncial source of what the synth's settings are.
This is to have an easily accessible, clear collection of the synths
data so it doesn't need to be fetched from the DOM or the web audio
nodes.