Use a dedicated constant source node to control voice3 detuning
An oscillators detune parameter can accept many inputs and it sums them
together, so instead of keeping track of a coefficient and recalculating
the oscillator frequency with it, just use a constant source node and
set the value (called offset) in cents, then connect it to the
oscillator detune.  This will allow oscillator 3 to go up an octave
(1200cents), up a fifth (700), or back to normal when the voice changes.