Add basic delay functionality to keys emulator
The web audio API has it's own DelayNode so I gave it a shot.  Between
the filter node and the master amp I placed a delay node, another filter
because the delay is very low rez, and an amp, all in series, that will
feed into the master amp and back into the delay.  The same amp controls
delay feedback and delay mix.  This is how it works on the actual synth.