Add Unison Ring voice to keys emulator
IMPORTANT: Look at the node diagram in /doc.

Ring modulation is when the gain of one oscillator is modulated by
another oscillator, with the range of 1 to -1.  This change adds in six
gainNodes that act as three switches, one for each oscillator.  All
three "switches" are governed by contant source nodes that toggle their
volumes.  The switches change signal path of each oscillator to go
through to the ampEg for unison mode, or to carry/modulate gain in a
three oscillator ring mod setup.