Allow octave change during polyphony
This involved doing away with the audio engine's notePlaying variable
which was used a lot for determining frequency and note changes, etc.
Instead, the core emulator code just sends the note as an argument to
the audio engine action.  Then it's set in the oscillatorNoteMap which
makes it easier to manage polyphony.