Add VCO EG INT functionality to keys emulator
This makes the detune value of the oscillators follow the envelope
generator.  This was tricky because on the actualy synth, the sustain
level is always non-detuned (detune = 0), so it wasn't as easy as
feeding the envelope to the detune and that's it.  The envelope is run
through a wave shaper, whose curve is normalized so that sustain = 0,
every time the sustain knob is changed.  The output of the wave shaper
is then amplified so that the value is significant for detune change in
cents.

What made this tricky is even though the envelope only outputs value
between 0 and 1, the wave shaper operates on a range between -1 and 1,
and I suspect it remaps the values somehow.  I had to tinker a lot to
find out that the curve should work within the -1 to 1 range.  A note on
debugging:  Since it is not easy to monitor raw signal levels in web
audio API, when I was debugging, I controlled the envelope value
directly with a slider instead of the normal user actions like
playNewNote etc.  This was a tough one!  Glad I figured it out though.