Switch ring mod paths on the fly in poly ring voice
In poly ring voice, the actual synth switches which oscillators are
carriers and which are modulating.  I observed on the synth that
oscillator playing the lowest note is always the carrier.  Up until this
point, the emulation always had osc1 as the carrier and then 2 and 3 as
subsequent modulators.  Rather than swap which oscillator was playing
what note to satisfy the lowest-note-is-always-carrier requirement, I
decided to keep each oscillators note and change the signal path
dynamically.  This is corroborated by the fact that detune on the main
synth is oscillator specific and switching the oscillators note instead
of the signal path would cause detune to behave differently.

The implementation was actually easier than I expected it to be.
Instead of fixed signal path per oscillator like before, now each
oscillator is connected to an amp that can send its audio signal to the
carrier or the two modulator destination, and each is controlled by
tuning a gainNode on or off.

See doc/keys_emulator_poly_ring_mod_dynamic_signal_path.txt for a
diagram.

When a new poly note plays in the poly ring voice, this is how the
signal path is determined:
 - Determine order of notes from lowest to highest
 - Create an array of signal path gainNodes in carrier, gain2 and gain3
   order.
 - For each oscillator, get its place in in the order of notes, and
   enable the corresponding gainNode.