Add modal to callout keys emulator
This is direct users to the new feature.  It only shows if the it's
enabled via an environment variable called
FEATURE_ENABLED_KEYS_EMULATOR_CALLOUT.  When the users sees it for the
first time, a cookie is set so that they won't see it again.  The modal
does not appear if the user landed on the keys emulator page first.
Google analytics events are fired depending on the action the user took
at the modal (click through or dismiss).  The GA is mainly because I
don't do modals and I'm curious how people interact with them.  I don't
want to annoy users so if I see high dismissal rate or bounce rate I'll
adjust my approach in the future.