Allow octave change using knob on keys emulator
One of the behaviors carried over from the bass emulator to the keys was
that Z and X had to be used to change octave.  I don't recall exactly
why (maybe it's written in the git log somewhere) but I assume it's
because the octaves change incrementally and I didn't know how to
suppress the knobturn event in between increments, which might have
caused performance issues.

This changes uses the suppressed knobturn events for snap knobs so now
the octave can be changed using the knob on the keys emulator interface
instead of just the Z and X keys, without worrying about performance.

An added benefit is that its better UX and I've alleviated the need for
the tooltip to show users how to change octave.