Add MIDI input functionality to keys emulator
This allows user to control the keys emulator using an external midi
controller, in addition to the existing inputs.  If the user is on
Chrome or other browser that supports Web MIDI, and there'e a MIDI
device available, a section is shown that let's the user choose their
input device and channel.  These can be changed as needed.  I tested
this personally with an AKAI MPK mini and a Roland TB-3 as MIDI inputs.

One new capability is using the sequencer or arpeggiator of the MIDI
controller to play sequences, which the emulator doesn't currently have.

I also assume you can open two tabs with two separate emulator instances
and play two different patches with two different midi controllers
simultaneously.