Migrate bass emulator to full Tone.js components
Ever since the sequencer was added, it had problems on Chrome.  I
investigated it for several hours years ago and eventually gave up, but
it seemed like the scheduling of the sequencer loop was outpacing the
scheduling of the playing of each note.  I don't know why the clocks of
each were out of sync, and I won't go into detail about what
specifically I found during debugging.  But I was unable to fix it.

Revisiting this, I decided to migrate all the vanilla web audio nodes to
their Tone.js counterparts, with the hope of ensuring all Tone
components would be aligned to the same clock.  With a heavy-handed
find-and-replace, then see what breaks approach, I was able to replace
all the web audio nodes with tone in about 30 minutes and the problem is
solved.  Furthermore I see no regressions in the non-sequencer
functionality, most likely because Tone components implement many of the
same functions as the underlying web audio nodes.

The only notable changes are this:
 - Constant source nodes have been replaces by Tone.Signal
 - Some places where I relied on Tone.Param to manage web audio node
   params have been removed, because now I can just manage Tone params
   directly.

Interface elements alerting the user that the sequencer is experimental,
or has compatibility issues, have been removed.

This change was way easier than I thought and I wish I took a crack at
it sooner.  Really glad it works though.