Migrate keys emulator to full Tone.js components
With the introduction of the keys sequencer, Tone is needed to keep
timing consistent.  This also simplifies some things.

Most of this change was simple find and replaces:
audioCtx.currentTime => Tone.now()
audioCtx.createGain() => new Tone.Gain()
and so on...

Notably, offset params aren't needed anymore because native
ConstantSource nodes are converted to Tone.Signal() objects.  So we can
work with the signal object directly instead of an offset param.  Signal
nodes also don't need to be started using start();