Force note release when running keys sequencer
Without this, the triggerDecayRelease function would calculate the end
of the note as the beginning of the release.  This is because the
oscillator notes are realtime values and can't be scheduled to change,
so we're not able to schedule them to be set to negative one in the
future to properly schedule the release.

The solution to this is to force the release when the sequencer is
running and don't even look at the oscillator notes if this is the case.