Fix snap knob from bugging to highest value
This was due to the fact that when the knob turn ends, a final
calculateDegree happens and at this point, the mouse is no longer
dragging, so the function would apply midi input control logic, even
though there was none.  This resulted in a huge difference, causing the
knobs rotation to go to its rightmost limit instead of stay put.